Add new Kubernetes Manifest chart
A new Helm chart named "Manifest" has been introduced to create generic Kubernetes resources. This new chart is specifically meant to be used when there are certain limitations with the conventional terraform manifest and the Kubernetes cluster must be present in the infrastructure beforehand. The chart includes a values.yaml file defining an example ConfigMap and Deployment, and a templates resources.yaml file. A maintainer information with the version and necessary metadata for the chart has been also provided.